Began implementing better water creation logic.
Water blocks are now created based on the height of the block currently
"in focus". A terrain map is generated based on the heights of each of
the topmost blocks instantiated in the game, and each block of a height
equal or less than the height of the in-focus block gets flooded.

Some fixes still need to be made to make sure a generated pool of water
blocks is fully connected, and that proper care is taken regarding earth
blocks adjacent to water blocks getting lowered.