Fix tutorial visualization
We now see the tutorial correctly when VR is off and when it is on. To
fix the problem with the water (which is a 3D object), we added a
camera for each tutorial panel, which outputs to a RenderTexture, which
is then shown on a RawImage in the middle of the panel.